# performance improvement: avoid superfluous `attr_accessor` creation